Change chat id type to Long

+++ src/main/java/ru/romangr/catbot/telegram/TelegramAdminNotifier.kt
@@ -50,5 +50,5 @@ class TelegramAdminNotifier(private val actionFactory: TelegramActionFactory,
     }
 
     private fun sendMessageAction(it: Long, message: String) =
-            actionFactory.newSendMessageAction(Chat(it.toInt()), message)
+            actionFactory.newSendMessageAction(Chat(it), message)
 }
